+++ packages/sst/src/constructs/Function.ts
@@ -72,6 +72,7 @@ const supportedRuntimes = {
   "nodejs18.x": CDKRuntime.NODEJS_18_X,
   "nodejs20.x": CDKRuntime.NODEJS_20_X,
   "nodejs22.x": CDKRuntime.NODEJS_22_X,
+  "nodejs24.x": CDKRuntime.NODEJS_24_X,
   "python3.7": CDKRuntime.PYTHON_3_7,
   "python3.8": CDKRuntime.PYTHON_3_8,
   "python3.9": CDKRuntime.PYTHON_3_9,
@@ -229,7 +230,7 @@ export interface FunctionProps
   handler?: string;
   /**
    * The runtime environment for the function.
-   * @default "nodejs22.x"
+   * @default "nodejs24.x"
    * @example
    * ```js
    * new Function(stack, "Function", {
@@ -881,7 +882,7 @@ export class Function extends CDKFunction implements SSTConstruct {
       .forEach((per) => {
         props = Function.mergeProps(per, props);
       });
-    props.runtime = props.runtime || "nodejs22.x";
+    props.runtime = props.runtime || "nodejs24.x";
     if (props.runtime === "go1.x") useWarning().add("go.deprecated");
 
     // Set defaults